marketplace-feature/marketplace-data-imp/src/main/kotlin/com/ilizma/marketplace/data/mapper/DiscountMapper.kt
package com.ilizma.marketplace.data.mapper

import com.ilizma.marketplace.domain.model.Discount
import com.ilizma.marketplace.domain.model.Discount.*
import com.ilizma.marketplace.data.model.Discount as DataDiscount

class DiscountMapper {

    fun from(
        product: DataDiscount,
    ): Discount = when (product) {
        is DataDiscount.TShirt -> TShirt(
            description = product.description,
        )
        is DataDiscount.Voucher -> Voucher(
            description = product.description,
        )
    }

}